Move rs_matter::codec to rs_matter::utils::codec as it is a generic base38 codec that does not depend on the rest of rs_matter; rename and move utils/mod.rs to utils.rs
