--- rs-matter/src/codec/mod.rs
@@ -1,1 +1,0 @@
-pub mod base38;
