+++ rs-matter/src/utils/codec/base38.rs
